backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9
python-dotenv>=1.0
django-cors-headers>=4.0
requests>=2.31

# Production & Payment
stripe>=7.0.0
sentry-sdk>=1.40.0

# Redis & Caching
redis>=5.0.0
django-redis>=5.4.0

# File Processing
pillow>=10.0.0
openpyxl>=3.1.0
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Production Server
gunicorn>=21.2.0

# WebSocket support
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0